refactor(SearchRanking): Execute Phase 1 & 2 (#VIG-289 / refactor-plan §3.4)
