+++ Packages/Package.swift
@@ -29,6 +29,7 @@ let macOSOnlyProducts: [Product] = [
     .singleTargetLibrary("Core"),
     .singleTargetLibrary("Cleanup"),
     .singleTargetLibrary("Search"),
+    .singleTargetLibrary("SearchRanking"),
     .singleTargetLibrary("SampleIndex"),
     .singleTargetLibrary("Services"),
     .singleTargetLibrary("Distribution"),
@@ -271,11 +272,20 @@ let targets: [Target] = {
         // the Strategies/ folder moves to Sources/SearchStrategies/ and gets its own
         // SPM target with deps: [SearchIndexCore, CoreJSONParser, CorePackageIndexing,
         // Core, SharedModels, SharedConstants, Resources, Logging].
-        dependencies: ["SharedCore", "SharedConstants", "SharedModels", "Logging", "CoreProtocols", "CoreJSONParser", "CorePackageIndexing", "Core", "ASTIndexer"]
+        dependencies: ["SharedCore", "SharedConstants", "SharedModels", "Logging", "CoreProtocols", "CoreJSONParser", "CorePackageIndexing", "Core", "ASTIndexer", "SearchRanking"]
     )
     let searchTestsTarget = Target.testTarget(
         name: "SearchTests",
         dependencies: ["Search", "SharedCore", "SharedConstants", "SharedModels", "SharedUtils", "TestSupport", "CorePackageIndexing"]
+    )
+
+    let searchRankingTarget = Target.target(
+        name: "SearchRanking",
+        dependencies: ["SharedCore", "SharedConstants", "SharedModels", "Logging"]
+    )
+    let searchRankingTestsTarget = Target.testTarget(
+        name: "SearchRankingTests",
+        dependencies: ["SearchRanking", "TestSupport"]
     )
 
     let sampleIndexTarget = Target.target(
@@ -542,6 +552,8 @@ let targets: [Target] = {
         cleanupTestsTarget,
         searchTarget,
         searchTestsTarget,
+        searchRankingTarget,
+        searchRankingTestsTarget,
         sampleIndexTarget,
         sampleIndexTestsTarget,
         servicesTarget,
